+++ polyclonal/polyclonal.py
@@ -661,7 +661,10 @@ class Polyclonal:
       self_initial_epitope self_harmonized_epitope ref_epitope  correlation
     0                   e1                      e1          e1          1.0
     1                   e2                      e2          e2          1.0
-    >>> assert model.mut_escape_df.equals(model_harmonized.mut_escape_df)
+    >>> if not model.mut_escape_df.equals(model_harmonized.mut_escape_df):
+    ...     raise ValueError(
+    ...         f"{model.mut_escape_df=}\n{model_harmonized.mut_escape_df=}"
+    ...     )
 
     >>> inverted_harmonized, harmonize_df = inverted_model.epitope_harmonized_model(
     ...     ref_model
@@ -3259,6 +3262,10 @@ class Polyclonal:
                 f"cannot harmonize 1-to-1:\n{corr_df=}\n{harmonize_df=}"
             )
 
+        # if only one epitope, do not need to do anything more
+        if len(self.epitopes) == 1:
+            return copy.deepcopy(self), harmonize_df
+
         map_dict = harmonize_df.set_index("self_initial_epitope")[
             "self_harmonized_epitope"
         ].to_dict()
@@ -3297,7 +3304,7 @@ class Polyclonal:
             alphabet=self.alphabet,
             epitope_colors=ref_poly.epitope_colors,
             data_mut_escape_overlap="exact_match",  # should be exact match in self
-            sites=self.sites,
+            sites=None if self.sequential_integer_sites else self.sites,
         )
         assert ref_poly.epitopes == harmonized_model.epitopes
         return harmonized_model, harmonize_df